add test test_loop_redirection.py, and api_get_short_url.py
