+++ API/TEST/test_create_short_url.py
@@ -9,7 +9,6 @@ from API.DATA.loopback_url import LOOPBACK_URL
 from API.FRAMEWORK.api_endpoints.api_short_link import ShorteningLinkAPI
 from API.FRAMEWORK.assertion.assert_content_type import assert_content_type
 from API.FRAMEWORK.assertion.assert_status_code import assert_status_code
-from API.FRAMEWORK.tools.create_short_url import create_short_url
 
 
 @allure.feature("Short URL generation")
